requirements-dev.txt
aegis-os==0.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.4
cryptography==46.0.5
distlib==0.4.0
docker==7.1.0
ecdsa==0.19.1
fastapi==0.135.1
filelock==3.25.0
fqdn==1.5.1
freezegun==1.5.5
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hatchling==1.29.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.17
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isoduration==20.11.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.3.1
librt==0.8.1
mypy==1.19.1
mypy_extensions==1.1.0
nexus-rpc==1.3.0
nodeenv==1.10.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pathspec==1.0.4
pip==26.0.1
platformdirs==4.9.2
pluggy==1.6.0
pre_commit==4.5.1
prometheus_client==0.24.1
protobuf==6.33.5
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-discovery==1.1.0
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.4
setuptools==80.10.2
six==1.17.0
starlette==0.52.1
structlog==25.5.0
temporalio==1.23.0
testcontainers==4.14.1
trove-classifiers==2026.1.14.14
types-protobuf==6.32.1.20260221
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
virtualenv==21.1.0
watchfiles==1.1.1
webcolors==25.10.0
websockets==16.0
wheel==0.46.3
wrapt==2.1.1
zipp==3.23.0
